component structure changed
